+++ Samples/pyRevit/Extensions/duHast-2025.extension/duHast.tab/lib/pushIt_associated/area_by_room_verification/update_pushit_rooms.py
@@ -20,9 +20,9 @@
 #
 #
 
+from System import Int64 # revit element Id expects 64 bit integer
 
-
-from duHast.Revit.DetailItems.curve_create import draw_2D_lines_on_bounding_box, draw_2D_lines_on_bounding_box_and_separate_point
+#from duHast.Revit.DetailItems.curve_create import draw_2D_lines_on_bounding_box, draw_2D_lines_on_bounding_box_and_separate_point
 
 
 from duHast.Utilities.Objects.result import Result
@@ -186,7 +186,7 @@ def update_push_it_instance(doc, push_it_family_instance,room):
             print("Room perimeter: {}".format(room_perimeter))
 
         #get the wall thickness parameter value from the push it family instance
-        half_wall_thickness_parameter_value = get_parameter_value_by_name(doc.GetElement(ElementId(push_it_family_instance.revit_element_id_integer_value)).Symbol , "HSL_WALL_THICKNESS", getter_double_as_double_converted_to_metric)
+        half_wall_thickness_parameter_value = get_parameter_value_by_name(doc.GetElement(ElementId(Int64(push_it_family_instance.revit_element_id_integer_value))).Symbol , "HSL_WALL_THICKNESS", getter_double_as_double_converted_to_metric)
 
         # make sure there is a fallback if no parameter is set
         half_wall_thickness = half_wall_thickness_parameter_value/2 /1000 if half_wall_thickness_parameter_value is not None else 0.06 # default to 60mm if not set
@@ -198,7 +198,7 @@ def update_push_it_instance(doc, push_it_family_instance,room):
         print("Calculated area: {}".format(calculated_area))
 
          # get the push it instance area stored in Area_Calc ( if that parameter does not exist, pop message and move on )
-        push_it_area = get_parameter_value_by_name(doc.GetElement(ElementId(push_it_family_instance.revit_element_id_integer_value)) , "Area_Calc", getter_double_as_double_converted_to_metric)
+        push_it_area = get_parameter_value_by_name(doc.GetElement(ElementId(Int64(push_it_family_instance.revit_element_id_integer_value))) , "Area_Calc", getter_double_as_double_converted_to_metric)
         print("Push it area: {}".format(push_it_area))
 
         if push_it_area is None:
@@ -213,7 +213,7 @@ def update_push_it_instance(doc, push_it_family_instance,room):
         if calculated_area < ( push_it_area * 0.8):
             return_value.update_sep(False, "Calculated area {} is less than 80% of push it area {}. Cannot update push it instance {} area.".format(calculated_area, push_it_area, push_it_family_instance.get_ui_name()))
             print_error("Calculated area {} is less than 80% of push it area {}. Cannot update push it instance {} area.".format(calculated_area, push_it_area, push_it_family_instance.get_ui_name()))
-            set_area_override_result = set_parameter_value_by_name(doc.GetElement(ElementId(push_it_family_instance.revit_element_id_integer_value)), "HSL_AREA_BY_REVIT_ROOM", "0.0")
+            set_area_override_result = set_parameter_value_by_name(doc.GetElement(ElementId(Int64(push_it_family_instance.revit_element_id_integer_value))), "HSL_AREA_BY_REVIT_ROOM", "0.0")
             print("Setting push it instance {} area override to 0.0".format(push_it_family_instance.get_ui_name()))
             print(set_area_override_result.message)
             return_value.update(set_area_override_result)
@@ -221,14 +221,14 @@ def update_push_it_instance(doc, push_it_family_instance,room):
         if calculated_area > ( push_it_area * 1.05):
             return_value.append_message("Calculated area {} is greater then push it area {}. Will not update push it instance {} area.".format(calculated_area,  push_it_area, push_it_family_instance.get_ui_name()))
             print("Calculated area {} is greater then push it area {}. Will not update push it instance {} area.".format(calculated_area,  push_it_area, push_it_family_instance.get_ui_name()))
-            set_area_override_result = set_parameter_value_by_name(doc.GetElement(ElementId(push_it_family_instance.revit_element_id_integer_value)), "HSL_AREA_BY_REVIT_ROOM", "0.0")
+            set_area_override_result = set_parameter_value_by_name(doc.GetElement(ElementId(Int64(push_it_family_instance.revit_element_id_integer_value))), "HSL_AREA_BY_REVIT_ROOM", "0.0")
             print("Setting push it instance {} area override to 0.0".format(push_it_family_instance.get_ui_name()))
             print(set_area_override_result.message)
             return_value.update(set_area_override_result)
             return return_value
 
         # get the push it override area
-        push_it_override_area = get_parameter_value_by_name(doc.GetElement(ElementId(push_it_family_instance.revit_element_id_integer_value)) , "HSL_AREA_BY_REVIT_ROOM", getter_double_as_double_converted_to_metric)
+        push_it_override_area = get_parameter_value_by_name(doc.GetElement(ElementId(Int64(push_it_family_instance.revit_element_id_integer_value))) , "HSL_AREA_BY_REVIT_ROOM", getter_double_as_double_converted_to_metric)
         print("Push it override area: {}".format(push_it_override_area))
         
         # compare to one decimal place, if identical do not take any action
@@ -239,13 +239,13 @@ def update_push_it_instance(doc, push_it_family_instance,room):
             # check if a previous override needs re-setting
             if push_it_override_area is not None and push_it_override_area != 0.0:
                 print("Push it override area needs updating from {} to 0.0".format(push_it_override_area))
-                set_area_override_result = set_parameter_value_by_name(doc.GetElement(ElementId(push_it_family_instance.revit_element_id_integer_value)), "HSL_AREA_BY_REVIT_ROOM", "0.0")
+                set_area_override_result = set_parameter_value_by_name(doc.GetElement(ElementId(Int64(push_it_family_instance.revit_element_id_integer_value))), "HSL_AREA_BY_REVIT_ROOM", "0.0")
                 return_value.update(set_area_override_result)
             return return_value
         
         print("Updating push it instance area to: {} from {}".format(calculated_area,  push_it_area ))
 
-        set_result = set_parameter_value_by_name(doc.GetElement(ElementId(push_it_family_instance.revit_element_id_integer_value)), "HSL_AREA_BY_REVIT_ROOM", str(calculated_area))
+        set_result = set_parameter_value_by_name(doc.GetElement(ElementId(Int64(push_it_family_instance.revit_element_id_integer_value))), "HSL_AREA_BY_REVIT_ROOM", str(calculated_area))
         return_value.update(set_result)
 
         print(set_result.message)
