+++ Samples/pyRevit/Extensions/duHast-2025.extension/duHast.tab/lib/pushIt_associated/place_revit_rooms/debug.py
@@ -1,5 +1,5 @@
 
-
+from System import Int64 # revit element Id expects 64 bit integer
 
 from duHast.Revit.DetailItems.curve_create import draw_2D_lines_on_bounding_box_and_separate_point
 
@@ -12,7 +12,7 @@ from Autodesk.Revit.DB import ElementId, Options, Solid, XYZ
 def draw_bounding_box_and_centroid(doc, family_instances):
     for fam in family_instances:
 
-        revit_family_instance = doc.GetElement(ElementId(fam.revit_element_id_integer_value))
+        revit_family_instance = doc.GetElement(ElementId(Int64(fam.revit_element_id_integer_value)))
         # get the bounding box of the family instance
         bounding_box = revit_family_instance.get_BoundingBox(doc.ActiveView)
         if bounding_box is not None:
@@ -32,7 +32,7 @@ def draw_bounding_box_and_centroid(doc, family_instances):
 def draw_bounding_box_and_centroid_linked_model(doc_host, doc_link, family_instances):
     for fam in family_instances:
 
-        revit_family_instance = doc_link.GetElement(ElementId(fam.revit_element_id_integer_value))
+        revit_family_instance = doc_link.GetElement(ElementId(Int64(fam.revit_element_id_integer_value)))
         # get the bounding box of the family instance
         bounding_box = revit_family_instance.get_BoundingBox(None)
         if bounding_box is not None:
